Fix Power of two Now output contains int values only

+++ src/main/java/school/mjc/stage0/loops/task2/PowerOfTwo.java
@@ -2,7 +2,6 @@ package school.mjc.stage0.loops.task2;
 
 public class PowerOfTwo {
     public void printPower(int power) {
-        // 0
         boolean isNegative = power < 0;
         int counter = 0;
 
@@ -10,7 +9,7 @@ public class PowerOfTwo {
             System.out.println("too much power");
         } else {
             while(counter <= power) {
-                System.out.println(Math.pow(2, counter));
+                System.out.println((int)Math.pow(2, counter));
                 counter++;
             }
         }
